chore: move scripts to bin
